Move cart initialization logic out of Layout component
- Create dedicated CartInitializer component for cart state management
- Move cart initialization to app root level in _app.tsx
- Remove cart-related code from Layout component

This change improves separation of concerns and follows DRY principles by
centralizing cart state initialization in a single location. The Layout
component is now focused solely on page structure, making the codebase
more maintainable.
