+++ src/components/Cart/CartInitializer.component.tsx
@@ -1,0 +1,43 @@
+import { useEffect } from 'react';
+import { useQuery } from '@apollo/client';
+
+// State
+import { useCartStore } from '@/stores/cartStore';
+
+// Utils
+import { getFormattedCart } from '@/utils/functions/functions';
+
+// GraphQL
+import { GET_CART } from '@/utils/gql/GQL_QUERIES';
+
+/**
+ * Non-rendering component responsible for initializing the cart state
+ * by fetching data from WooCommerce and syncing it with the Zustand store.
+ * This should be rendered once at the application root (_app.tsx).
+ * @function CartInitializer
+ * @returns {null} - This component does not render any UI
+ */
+const CartInitializer = () => {
+  const { syncWithWooCommerce } = useCartStore();
+
+  const { data, refetch } = useQuery(GET_CART, {
+    notifyOnNetworkStatusChange: true,
+    onCompleted: () => {
+      // On successful fetch, format the data and sync with the store
+      const updatedCart = getFormattedCart(data);
+      if (updatedCart) {
+        syncWithWooCommerce(updatedCart);
+      }
+    },
+    // Consider error handling if needed (e.g., onError callback)
+  });
+
+  useEffect(() => {
+    refetch();
+  }, [refetch]);
+
+  // This component does not render any UI
+  return null;
+};
+
+export default CartInitializer;
